+++ Example/VimeoUpload-iOS-Example/VimeoUpload-iOS-Example/AppDelegate.swift
@@ -49,14 +49,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate
         
         UploadManager.sharedInstance.applicationDidFinishLaunching() // Ensure init is called on launch
         
-        UploadManager.sharedInstance.reporter.sendMessage("Launch")
-        
         return true
     }
 
     func application(application: UIApplication, handleEventsForBackgroundURLSession identifier: String, completionHandler: () -> Void)
     {
-        if UploadManager.sharedInstance.descriptorManager.handleEventsForBackgroundURLSession(identifier, completionHandler: completionHandler) == false
+        if UploadManager.sharedInstance.handleEventsForBackgroundURLSession(identifier: identifier, completionHandler: completionHandler) == false
         {
             assertionFailure("Unhandled background events")
         }
