--- Example/VimeoUpload-iOS-Example/VimeoUpload-iOS-Example/ViewControllers/CameraRollViewController.xib
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="9059" systemVersion="14F27" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9049"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CameraRollViewController" customModule="VimeoUpload_iOS_2Step" customModuleProvider="target">
-            <connections>
-                <outlet property="activityIndicatorView" destination="KJx-Rw-dOW" id="bJX-QE-Alz"/>
-                <outlet property="collectionView" destination="hCc-kg-2a0" id="ZhC-oC-ffj"/>
-                <outlet property="view" destination="i5M-Pr-FkT" id="YcC-eA-0Vq"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
-            <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="hCc-kg-2a0">
-                    <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="ex2-yS-faZ">
-                        <size key="itemSize" width="50" height="50"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                    <connections>
-                        <outlet property="dataSource" destination="-1" id="suP-tl-XT1"/>
-                        <outlet property="delegate" destination="-1" id="FzO-Yk-Bb7"/>
-                    </connections>
-                </collectionView>
-                <activityIndicatorView hidden="YES" opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" hidesWhenStopped="YES" style="whiteLarge" translatesAutoresizingMaskIntoConstraints="NO" id="KJx-Rw-dOW">
-                    <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                    <color key="backgroundColor" red="0.0" green="0.47843137250000001" blue="1" alpha="0.5" colorSpace="calibratedRGB"/>
-                </activityIndicatorView>
-            </subviews>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
-            <constraints>
-                <constraint firstItem="KJx-Rw-dOW" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="6BX-C4-6vX"/>
-                <constraint firstAttribute="trailing" secondItem="hCc-kg-2a0" secondAttribute="trailing" id="Cyf-wD-VlU"/>
-                <constraint firstItem="KJx-Rw-dOW" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="Kwl-Fb-E0q"/>
-                <constraint firstAttribute="bottom" secondItem="KJx-Rw-dOW" secondAttribute="bottom" id="L2I-RO-pSA"/>
-                <constraint firstItem="hCc-kg-2a0" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="TJ9-px-Sft"/>
-                <constraint firstItem="hCc-kg-2a0" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="U3v-8c-uRx"/>
-                <constraint firstAttribute="trailing" secondItem="KJx-Rw-dOW" secondAttribute="trailing" id="exf-Ve-z6v"/>
-                <constraint firstAttribute="bottom" secondItem="hCc-kg-2a0" secondAttribute="bottom" id="xOk-b8-KIP"/>
-            </constraints>
-        </view>
-    </objects>
-</document>
