+++ VimeoUpload/Controllers/ConnectivityManager.swift
